+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -270,7 +270,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfctasr55g1dr3xaapjkf0s4"),
+  _id: z.string().default("conv_01kfdc5fajhxqabwnsh3100j6v"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -302,7 +302,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfctasr55g1dr3xaapjkf0s4"),
+  id: z.string().default("conv_01kfdc5fajhxqabwnsh3100j6v"),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),