+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4703,7 +4703,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KFCTASPA00ETS7AA29KY8959"),
+  id: z.string().default("01KFDC5F8F38WG0ES1DHGH5HNY"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>
@@ -4724,7 +4724,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KFCTASPA00ETS7AA29KY8959"),
+  id: z.string().default("01KFDC5F8F38WG0ES1DHGH5HNY"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>